pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "waterbase"
version = "0.1.0"
description = ""
authors = ["apathriel <gabrielandersen1@gmail.com>", "missingusername"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.6"
notebook = "^7.2.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
url-normalize = "^1.4.3"
readability = "^0.3.1"
asyncpg = "^0.30.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
lxml = "^5.3.0"
langchain-community = "^0.3.9"
langchain-postgres = "^0.0.12"
psycopg = { extras = ["binary"], version = "^3.2.3" }
langgraph-checkpoint-postgres = "^2.0.7"
langgraph-cli = "^0.1.61"
streamlit = "^1.40.2"
pydantic = "^2.10.3"
streamlit-chat = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
isort = "^5.13.2"

[tool.poetry.group.web-scraping.dependencies]
requests = "^2.32.3"
playwright = "^1.48.0"

[tool.poetry.group.ai.dependencies]
langgraph = "^0.2.53"
langchain-openai = "^0.2.10"
langchain = "^0.3.8"
langserve = { extras = ["all"], version = "^0.3.0" }


[tool.poetry.group.database.dependencies]
sqlalchemy = "^2.0.36"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
